+++ packages/moonstone/CHANGELOG.md
@@ -16,7 +16,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Added
 
-- `ui/VirtualList.VirtualList` and `ui/VirtualList.VirtualGridList` support `data-webos-voice-focused` and `data-webos-voice-group-label`
+- `moonstone/VirtualList.VirtualList` and `moonstone/VirtualList.VirtualGridList` support `data-webos-voice-focused` and `data-webos-voice-group-label`
 
 ## [2.0.0-beta.9] - 2018-07-02